Initial commit implementing basic functionality
This is a rought first draft the code needs much more structure. At
present the code make use of the axios hook library to fetch the data
and display it in a grid. In addition a row can be updated when the user
clicks on a row. Added some basics tests as well to test this
functionality.