Adds filtering, cleans up fetching
This adds a some filtering in the form of checkboxes. This can be used
to filter the different statuses of the tubes.

In addition to these changes, a testing and types folder has been added.
The types folder is to store different types that will be share across
components. The testing folder will store utility functions for testing.
This include testData generate. This will help aid maintainability of
the tests.